Improved boot scripts

+++ stopall.sh
@@ -1,5 +1,5 @@
 docker stop $(docker ps -a -q)
 
-if [ $1 = "-d" ]; then
+if [ $# -ge 1 ] && [ $1 = "-d" ]; then
 	docker rm $(docker ps -a -q)
 fi
